refactor: 상위 layout이 prefetch 하는 data 정보를 UserRegion 에서 MyInfo 로 수정 (#206)
