+++ apps/web/lib/query/server/index.ts
@@ -1,2 +1,2 @@
 export { createServerQueryClient } from './instance';
-export * from './prefetchUserQuery';
+export * from './prefetchMyInfo';
