+++ apps/web/services/my/client/index.ts
@@ -1,0 +1,1 @@
+export * from './fetchMyInfo';
