+++ apps/web/services/my/server/getMyInfo.ts
@@ -2,21 +2,25 @@
 
 import { prisma } from '@web/lib/prisma';
 
-import type { UserRegion } from '@web/types';
+import type { MyInfoAPIResponse } from '@web/types';
 import { getAuthenticatedUser } from '@web/utils/auth/server';
 
-export const getUserRegion = async (): Promise<UserRegion> => {
+export const getMyInfo = async (): Promise<MyInfoAPIResponse> => {
   try {
     const authResult = await getAuthenticatedUser();
 
     const user = await prisma.users.findUnique({
       where: { user_id: authResult.userId },
       select: {
+        user_id: true,
         region: true,
       },
     });
 
-    return { region: user?.region ?? '', userId: authResult.userId ?? '' }; // TODO: null 일 경우 어떻게 처리하는 것이 가장 좋을까?
+    return {
+      userId: user?.user_id ?? '',
+      region: user?.region ?? '',
+    }; // TODO: null 일 경우 어떻게 처리하는 것이 가장 좋을까?
   } catch (error) {
     if (process.env.NODE_ENV === 'development') {
       // eslint-disable-next-line no-console
