+++ apps/web/services/my/server/index.ts
@@ -1,0 +1,1 @@
+export * from './getMyInfo';
