--- apps/web/services/user/client/index.ts
@@ -1,1 +1,0 @@
-export * from './fetchUserRegion';
