--- apps/web/services/user/server/index.ts
@@ -1,1 +1,0 @@
-export * from './getUserRegion';
