+++ apps/web/types/index.ts
@@ -2,4 +2,4 @@ export type * from './location';
 export type * from './image';
 export type * from './product';
 export type * from './chat';
-export type * from './user';
+export type * from './my-info';
