+++ apps/web/types/my-info/api.ts
@@ -1,4 +1,4 @@
-export interface UserRegion {
+export interface MyInfoAPIResponse {
   region: string;
   userId: string;
 }
